Refactoring: move a search function to search.js
