+++ package.json
@@ -49,6 +49,7 @@
     "gulp-cli": "^1.4.0",
     "node-dev": "^3.1.3",
     "style-loader": "^0.13.2",
+    "urijs": "^1.19.0",
     "vinyl-source-stream": "^1.1.0",
     "webpack": "3.0.0",
     "webpack-dev-server": "^1.16.5",
