--- src/page/index.js
@@ -1,3 +1,0 @@
-/**
-    Functions for a PDFAnno web page.
-*/
